chore(ci,lint): harden lint pipeline and clean staged lint issues across meta/xmp/imagetype
- add `golangci.yml` and modernize `.github/workflows/golangci-lint.yml`
- update CI behavior for draft PRs, only-new-issues, and `govulncheck` security scanning
- tidy and refresh module metadata (`go.mod`/`go.sum`), including dependency updates and cleanup
- fix `imagetype` test lint failures (`errcheck`, `errorlint`, `govet`, `misspell`)
- fix `meta/*` lint classes requested: `govet`, `misspell`, `gofmt`, `wastedassign`, and `errorlint`
- refactor `meta/xmp` parsing helpers (`parseInt8`, safer pooled readers, `errors.Is` usage) and add targeted `gosec` suppressions for bounded casts

+++ meta/xmp/basic.go
@@ -24,7 +24,7 @@ func (basic *Basic) parse(p property) (err error) {
 	case ModifyDate:
 		basic.ModifyDate, err = parseDate(p.Value())
 	case Rating:
-		basic.Rating = int8(parseUint8(p.Value()))
+		basic.Rating = parseInt8(p.Value())
 	case XMPToolkit:
 		basic.Toolkit = parseString(p.Value())
 	default:
